Update the menu
