+++ src/App.tsx
@@ -7,13 +7,13 @@ import AccommodationPage from "./pages/Attend/Accommodation";
 import NoPage from "./pages/NoPage";
 import PresentPage from "./pages/Present";
 import AttendTGPPage from "./pages/Attend/TravelGrantProgram";
-import Workshops from "./pages/Workshops";
-import CommunityDayPage from "./pages/CommunityDay";
+import Workshops from "./pages/Program/Workshops";
+import CommunityDayPage from "./pages/Program/CommunityDay";
 import Sponsorship from "./pages/Sponsorship";
 import OurConferenceVenuePage from "./pages/Attend/OurConferenceVenue";
-import ProgramPage from "./pages/Attend/Program";
+import ProgramPage from "./pages/Program/Program";
 import RegisterPage from "./pages/Attend/Register";
-import SocialEventsPage from "./pages/Attend/SocialEvents";
+import SocialEventsPage from "./pages/Program/SocialEvents";
 import GetInvolvedPage from "./pages/Program/GetInvolved";
 
 export default function App() {
@@ -23,7 +23,7 @@ export default function App() {
         <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
           <Route path="code-of-conduct" element={<CodeOfConductPage />} />
-          <Route path="present" element={<PresentPage />} />
+          <Route path="attend/present" element={<PresentPage />} />
           <Route
             path="attend/travel-grant-program"
             element={<AttendTGPPage />}
@@ -33,16 +33,16 @@ export default function App() {
             element={<OurConferenceVenuePage />}
           />
           <Route path="attend/Accommodation" element={<AccommodationPage />} />
-          <Route path="attend/register" element={<RegisterPage />} />
-          <Route path="attend/social-events" element={<SocialEventsPage />} />
-          <Route path="workshops" element={<Workshops />} />
+          <Route path="attend" element={<RegisterPage />} />
+          <Route path="program/social-events" element={<SocialEventsPage />} />
+          <Route path="program/workshops" element={<Workshops />} />
           <Route path="program/get-involved" element={<GetInvolvedPage />} />
-          <Route path="community-day" element={<CommunityDayPage />} />
+          <Route path="program/community-day" element={<CommunityDayPage />} />
           <Route path="sponsorship" element={<Sponsorship />} />
           <Route path="*" element={<NoPage />} />
         </Route>
         <Route path="/" element={<LayoutNoFooter />}>
-          <Route path="attend/program" element={<ProgramPage />} />
+          <Route path="program" element={<ProgramPage />} />
         </Route>
       </Routes>
     </HashRouter>
