+++ src/documents/top-menu.md
@@ -1,13 +1,13 @@
 * [HOME](/)
-* [ATTEND](#/attend/register)
-    * [Register](#/attend/register)
-    * [Present](#/present)
+* [ATTEND](#/attend)
+    * [Register](#/attend)
+    * [Present](#/attend/present)
     * [Venue](#/attend/our-conference-venue)
     * [Accommodation](#/attend/Accommodation)
-* [PROGRAM](#/attend/program)
-    * [Schedule](#/attend/program)
-    * [Workshops](#/workshops)
-    * [Community Day](#/community-day)
-    * [Social Events](#/attend/social-events)
+* [PROGRAM](#/program)
+    * [Schedule](#/program)
+    * [Workshops](#/program/workshops)
+    * [Community Day](#/program/community-day)
+    * [Social Events](#/program/social-events)
     * [Get Involved](#/program/get-involved)
 * [SPONSORSHIP](#/sponsorship)
